Better handling of asserts, hard faults  (#52)
* add auto reboot and logging on serious faults
* hard fault
* assert
also clean up some dead weight code

Rationalize handling of 1 vs 2 FPGAs, code cleanup
* ensure that results for monitoring always return the same FPGA
* move some of the code out of cm_mcu.c into new file called LocalTask.c (can't be Task.c due to limitations of apple file system)

* change interface to error buffer to hide internals
also clean up command line task (get rid of extraneous variables)

* fix issues caught by clang that gcc did not catch
* error buffer string handling, clang makedefs

* * use EEPROM firefly info to avoid reading fireflies not populated

+++ common/smbus_units.c
@@ -13,7 +13,7 @@ float linear11_to_float(linear11_val_t t)
 
 float linear16u_to_float(uint16_t t )
 {
-  const float mantissa = -13;
+  const int mantissa = -13;
   return 1.0*t*pow(2,mantissa);
 }
